@@ -9,7 +9,7 @@
         }
     ],
     "require": {
-        "php": "8.*|7.4.*|7.3.*|7.2.*",
+        "php": "8.1",
         "ext-json": "*",
         "ext-iconv": "*",
         "ext-xmlwriter": "*",
@@ -17,7 +17,7 @@
         "ext-curl": "*",
         "ext-zip": "*",
         "ext-simplexml": "*",
-        "symfony/yaml": "6.*|5.*|^4.3|^3.4",
+        "symfony/yaml": "6.*",
         "beberlei/assert": "^3.3",
         "adhocore/cli": "^0.9.0",
         "firebase/php-jwt": "^v6.8",
