@@ -6,13 +6,7 @@ use Misery\Component\Action\ItemActionProcessor;
 
 class ActionPipe implements PipeInterface
 {
-    private $actionProcessor;
-
-    public function __construct(ItemActionProcessor $actionProcessor)
-    {
-
-        $this->actionProcessor = $actionProcessor;
-    }
+    public function __construct(private readonly ItemActionProcessor $actionProcessor) {}
 
     public function pipe(array $item): array
     {
