@@ -8,8 +8,9 @@ use Misery\Component\Configurator\ConfigurationAwareInterface;
 use Misery\Component\Configurator\ConfigurationTrait;
 use Misery\Component\Configurator\ReadOnlyConfiguration;
 use Misery\Component\Extension\ExtensionInterface;
+use Misery\Model\DataStructure\ItemInterface;
 
-class ExtensionAction implements OptionsInterface, ConfigurationAwareInterface
+class ExtensionAction implements OptionsInterface, ConfigurationAwareInterface, ActionItemInterface
 {
     use OptionsTrait;
     use ConfigurationTrait;
@@ -21,6 +22,21 @@ class ExtensionAction implements OptionsInterface, ConfigurationAwareInterface
     private $options = [
         'extension' => null,
     ];
+
+    public function applyAsItem(ItemInterface $item): ItemInterface
+    {
+        $extension = $this->getOption('extension');
+        if (null === $extension) {
+            return $item;
+        }
+        // loadExtension
+        if (null === $this->extension) {
+            $extensionFile = $this->configuration->getExtensions()[$extension.'.php'] ?? null;
+            $this->extension = $this->loadExtension($extensionFile, 'Extensions\\'.$extension);
+        }
+
+        return $this->extension->applyAsItem($item);
+    }
 
     public function apply($item): array
     {
